terrarium/requirements.txt
acme==1.1.0
asgiref==3.3.1
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certbot==0.40.0
certbot-apache==0.39.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Click==7.0
cloud-init==21.1
colorama==0.4.3
command-not-found==0.3
ConfigArgParse==0.13.0
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.7
dbus-python==1.2.16
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.7
django-allauth==0.44.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
entrypoints==0.3
future==0.18.2
gunicorn==20.1.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.10
importlib-metadata==1.5.0
incremental==16.10.1
install==1.3.4
iotop==0.6
Jinja2==2.10.1
josepy==1.2.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Markdown==3.3.4
MarkupSafe==1.1.0
mock==3.0.5
more-itertools==4.2.0
mysql-connector-python==8.0.23
mysqlclient==2.0.3
netifaces==0.10.4
oauthlib==3.1.0
paho-mqtt==1.5.1
parsedatetime==2.4
pbr==5.4.5
pexpect==4.6.0
protobuf==3.15.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
PyGObject==3.36.0
PyHamcrest==1.9.0
PyICU==2.4.2
PyJWT==2.0.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.4
python-augeas==0.5.0
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.15.0
sos==4.1
sqlparse==0.4.1
ssh-import-id==5.10
systemd-python==234
Twisted==18.9.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.3
urlpatterns==0.0.1
uWSGI==2.0.19.1
wadllib==1.3.3
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.0
zope.interface==4.7.1
